+++ openair2/RRC/NR_UE/L2_interface_ue.c
@@ -37,7 +37,7 @@
 
 typedef uint32_t channel_t;
 
-void nr_mac_rrc_meas_ind_ue(module_id_t module_id, uint32_t gNB_index, uint16_t Nid_cell, bool csi_meas, bool is_neighboring_cell, uint8_t rsrp_dBm)
+void nr_mac_rrc_meas_ind_ue(module_id_t module_id, uint32_t gNB_index, uint16_t Nid_cell, bool csi_meas, bool is_neighboring_cell, int rsrp_dBm)
 {
   MessageDef *message_p = itti_alloc_new_message(TASK_MAC_UE, 0, NR_RRC_MAC_MEAS_DATA_IND);
   NR_RRC_MAC_MEAS_DATA_IND(message_p).gnb_index = gNB_index;
